Rename buildIndex.tsx to Home.tsx
